move files to new src directory
