+++ test/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+      "outDir": "./built/test",
+      "allowJs": true,
+      "target": "es5"
+    },
+    "include": ["./test/**/*"]
+  }
